requirements.txt
Django>=4.2.18
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
python-environ==0.4.54